@@ -3,46 +3,46 @@
     [
       {
         "name": "react-addons-css-transition-group",
-        "raw": "react-addons-css-transition-group@15.3.2",
-        "rawSpec": "15.3.2",
+        "raw": "react-addons-css-transition-group@15.4.0",
+        "rawSpec": "15.4.0",
         "scope": null,
-        "spec": "15.3.2",
+        "spec": "15.4.0",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "react-addons-css-transition-group@15.3.2",
-  "_id": "react-addons-css-transition-group@15.3.2",
+  "_from": "react-addons-css-transition-group@15.4.0",
+  "_id": "react-addons-css-transition-group@15.4.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-addons-css-transition-group",
-  "_nodeVersion": "6.5.0",
+  "_nodeVersion": "6.9.1",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-addons-css-transition-group-15.3.2.tgz_1474307192205_0.6588692599907517"
+    "tmp": "tmp/react-addons-css-transition-group-15.4.0.tgz_1479306744579_0.6365639499854296"
   },
   "_npmUser": {
-    "email": "paul@oshannessy.com",
-    "name": "zpao"
+    "email": "tomocchino@gmail.com",
+    "name": "tomocchino"
   },
-  "_npmVersion": "3.10.3",
+  "_npmVersion": "3.10.8",
   "_phantomChildren": {},
   "_requested": {
     "name": "react-addons-css-transition-group",
-    "raw": "react-addons-css-transition-group@15.3.2",
-    "rawSpec": "15.3.2",
+    "raw": "react-addons-css-transition-group@15.4.0",
+    "rawSpec": "15.4.0",
     "scope": null,
-    "spec": "15.3.2",
+    "spec": "15.4.0",
     "type": "version"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/react-addons-css-transition-group/-/react-addons-css-transition-group-15.3.2.tgz",
-  "_shasum": "d8fa52bec9bb61bdfde8b9e4652b80297cbff667",
+  "_resolved": "https://registry.npmjs.org/react-addons-css-transition-group/-/react-addons-css-transition-group-15.4.0.tgz",
+  "_shasum": "410aa74533629974de33cfeaa85945066ff181cd",
   "_shrinkwrap": null,
-  "_spec": "react-addons-css-transition-group@15.3.2",
+  "_spec": "react-addons-css-transition-group@15.4.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
@@ -52,8 +52,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "d8fa52bec9bb61bdfde8b9e4652b80297cbff667",
-    "tarball": "https://registry.npmjs.org/react-addons-css-transition-group/-/react-addons-css-transition-group-15.3.2.tgz"
+    "shasum": "410aa74533629974de33cfeaa85945066ff181cd",
+    "tarball": "https://registry.npmjs.org/react-addons-css-transition-group/-/react-addons-css-transition-group-15.4.0.tgz"
   },
   "files": [
     "LICENSE",
@@ -90,6 +90,10 @@
       "name": "spicyj"
     },
     {
+      "email": "tomocchino@gmail.com",
+      "name": "tomocchino"
+    },
+    {
       "email": "paul@oshannessy.com",
       "name": "zpao"
     }
@@ -97,7 +101,7 @@
   "name": "react-addons-css-transition-group",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.3.2"
+    "react": "^15.4.0"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -105,5 +109,5 @@
     "url": "git+https://github.com/facebook/react.git"
   },
   "scripts": {},
-  "version": "15.3.2"
+  "version": "15.4.0"
 }
