@@ -3,47 +3,47 @@
     [
       {
         "name": "react-addons-shallow-compare",
-        "raw": "react-addons-shallow-compare@15.3.2",
-        "rawSpec": "15.3.2",
+        "raw": "react-addons-shallow-compare@15.4.0",
+        "rawSpec": "15.4.0",
         "scope": null,
-        "spec": "15.3.2",
+        "spec": "15.4.0",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "react-addons-shallow-compare@15.3.2",
-  "_id": "react-addons-shallow-compare@15.3.2",
+  "_from": "react-addons-shallow-compare@15.4.0",
+  "_id": "react-addons-shallow-compare@15.4.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-addons-shallow-compare",
-  "_nodeVersion": "6.5.0",
+  "_nodeVersion": "6.9.1",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-addons-shallow-compare-15.3.2.tgz_1474307206506_0.7146502370014787"
+    "tmp": "tmp/react-addons-shallow-compare-15.4.0.tgz_1479306757893_0.7042119661346078"
   },
   "_npmUser": {
-    "email": "paul@oshannessy.com",
-    "name": "zpao"
+    "email": "tomocchino@gmail.com",
+    "name": "tomocchino"
   },
-  "_npmVersion": "3.10.3",
+  "_npmVersion": "3.10.8",
   "_phantomChildren": {},
   "_requested": {
     "name": "react-addons-shallow-compare",
-    "raw": "react-addons-shallow-compare@15.3.2",
-    "rawSpec": "15.3.2",
+    "raw": "react-addons-shallow-compare@15.4.0",
+    "rawSpec": "15.4.0",
     "scope": null,
-    "spec": "15.3.2",
+    "spec": "15.4.0",
     "type": "version"
   },
   "_requiredBy": [
     "/",
     "/pui-react-mixins"
   ],
-  "_resolved": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.3.2.tgz",
-  "_shasum": "c9edba49b9eab44d0c59024d289beb1ab97318b5",
+  "_resolved": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.4.0.tgz",
+  "_shasum": "17b2cced1928937bbcdb4832fb200bc34bd47871",
   "_shrinkwrap": null,
-  "_spec": "react-addons-shallow-compare@15.3.2",
+  "_spec": "react-addons-shallow-compare@15.4.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
@@ -53,8 +53,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "c9edba49b9eab44d0c59024d289beb1ab97318b5",
-    "tarball": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.3.2.tgz"
+    "shasum": "17b2cced1928937bbcdb4832fb200bc34bd47871",
+    "tarball": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.4.0.tgz"
   },
   "files": [
     "LICENSE",
@@ -91,6 +91,10 @@
       "name": "spicyj"
     },
     {
+      "email": "tomocchino@gmail.com",
+      "name": "tomocchino"
+    },
+    {
       "email": "paul@oshannessy.com",
       "name": "zpao"
     }
@@ -98,7 +102,7 @@
   "name": "react-addons-shallow-compare",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.3.2"
+    "react": "^15.4.0"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -106,5 +110,5 @@
     "url": "git+https://github.com/facebook/react.git"
   },
   "scripts": {},
-  "version": "15.3.2"
+  "version": "15.4.0"
 }
