@@ -3,64 +3,80 @@
     [
       {
         "name": "react-dom",
-        "raw": "react-dom@15.3.2",
-        "rawSpec": "15.3.2",
+        "raw": "react-dom@15.4.0",
+        "rawSpec": "15.4.0",
         "scope": null,
-        "spec": "15.3.2",
+        "spec": "15.4.0",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "react-dom@15.3.2",
-  "_id": "react-dom@15.3.2",
+  "_from": "react-dom@15.4.0",
+  "_id": "react-dom@15.4.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-dom",
-  "_nodeVersion": "6.5.0",
+  "_nodeVersion": "6.9.1",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-dom-15.3.2.tgz_1474307217444_0.7416605225298554"
+    "host": "packages-18-east.internal.npmjs.com",
+    "tmp": "tmp/react-dom-15.4.0.tgz_1479306774757_0.8015152509324253"
   },
   "_npmUser": {
-    "email": "paul@oshannessy.com",
-    "name": "zpao"
+    "email": "tomocchino@gmail.com",
+    "name": "tomocchino"
   },
-  "_npmVersion": "3.10.3",
-  "_phantomChildren": {},
+  "_npmVersion": "3.10.8",
+  "_phantomChildren": {
+    "isomorphic-fetch": "2.2.1",
+    "loose-envify": "1.2.0",
+    "object-assign": "4.1.0",
+    "promise": "7.1.1",
+    "ua-parser-js": "0.7.12"
+  },
   "_requested": {
     "name": "react-dom",
-    "raw": "react-dom@15.3.2",
-    "rawSpec": "15.3.2",
+    "raw": "react-dom@15.4.0",
+    "rawSpec": "15.4.0",
     "scope": null,
-    "spec": "15.3.2",
+    "spec": "15.4.0",
     "type": "version"
   },
   "_requiredBy": [
     "/",
     "/react-bootstrap-table/react-toastr"
   ],
-  "_resolved": "https://registry.npmjs.org/react-dom/-/react-dom-15.3.2.tgz",
-  "_shasum": "c46b0aa5380d7b838e7a59c4a7beff2ed315531f",
+  "_resolved": "https://registry.npmjs.org/react-dom/-/react-dom-15.4.0.tgz",
+  "_shasum": "6a97a69000966570db48c746bc4b7b0ca50d1534",
   "_shrinkwrap": null,
-  "_spec": "react-dom@15.3.2",
+  "_spec": "react-dom@15.4.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
+  "browserify": {
+    "transform": [
+      "loose-envify"
+    ]
+  },
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
-  "dependencies": {},
+  "dependencies": {
+    "fbjs": "^0.8.1",
+    "loose-envify": "^1.1.0",
+    "object-assign": "^4.1.0"
+  },
   "description": "React package for working with the DOM.",
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "c46b0aa5380d7b838e7a59c4a7beff2ed315531f",
-    "tarball": "https://registry.npmjs.org/react-dom/-/react-dom-15.3.2.tgz"
+    "shasum": "6a97a69000966570db48c746bc4b7b0ca50d1534",
+    "tarball": "https://registry.npmjs.org/react-dom/-/react-dom-15.4.0.tgz"
   },
   "files": [
     "LICENSE",
     "PATENTS",
     "README.md",
     "dist/",
+    "lib/",
     "index.js",
     "server.js"
   ],
@@ -92,6 +108,10 @@
       "name": "spicyj"
     },
     {
+      "email": "tomocchino@gmail.com",
+      "name": "tomocchino"
+    },
+    {
       "email": "paul@oshannessy.com",
       "name": "zpao"
     }
@@ -99,7 +119,7 @@
   "name": "react-dom",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.3.2"
+    "react": "^15.4.0"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -109,5 +129,5 @@
   "scripts": {
     "start": "node server.js"
   },
-  "version": "15.3.2"
+  "version": "15.4.0"
 }
