@@ -6,7 +6,6 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  *
- * @providesModule LinkedStateMixin
  */
 
 'use strict';
@@ -24,8 +23,7 @@ var LinkedStateMixin = {
    * ReactLink will have the current value of this.state[key] and will call
    * setState() when a change is requested.
    *
-   * @param {string} key state key to update. Note: you may want to use keyOf()
-   * if you're using Google Closure Compiler advanced mode.
+   * @param {string} key state key to update.
    * @return {ReactLink} ReactLink instance linking to the state.
    */
   linkState: function (key) {
