@@ -6,13 +6,13 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  *
- * @providesModule ReactWithAddons
  */
 
 'use strict';
 
 var LinkedStateMixin = require('./LinkedStateMixin');
 var React = require('./React');
+var ReactAddonsDOMDependencies = require('./ReactAddonsDOMDependencies');
 var ReactComponentWithPureRenderMixin = require('./ReactComponentWithPureRenderMixin');
 var ReactCSSTransitionGroup = require('./ReactCSSTransitionGroup');
 var ReactFragment = require('./ReactFragment');
@@ -33,8 +33,20 @@ React.addons = {
 };
 
 if (process.env.NODE_ENV !== 'production') {
-  React.addons.Perf = require('./ReactPerf');
-  React.addons.TestUtils = require('./ReactTestUtils');
+  // For the UMD build we get these lazily from the global since they're tied
+  // to the DOM renderer and it hasn't loaded yet.
+  Object.defineProperty(React.addons, 'Perf', {
+    enumerable: true,
+    get: function () {
+      return ReactAddonsDOMDependencies.getReactPerf();
+    }
+  });
+  Object.defineProperty(React.addons, 'TestUtils', {
+    enumerable: true,
+    get: function () {
+      return ReactAddonsDOMDependencies.getReactTestUtils();
+    }
+  });
 }
 
 module.exports = React;
