@@ -1,4 +1,5 @@
 var assert      = require('assert');
+var requirejs   = require('requirejs');
 var UAParser    = require('./../src/ua-parser');
 var browsers    = require('./browser-test.json');
 var cpus        = require('./cpu-test.json');
@@ -43,15 +44,6 @@ describe('UAParser()', function () {
     assert.deepEqual(UAParser(ua), new UAParser().setUA(ua).getResult());
 });
 
-describe('Injected Browser', function () {
-    var uaString = 'ownbrowser/1.3';
-    var ownBrowser = [[/(ownbrowser)\/((\d+)?[\w\.]+)/i], [UAParser.BROWSER.NAME, UAParser.BROWSER.VERSION, UAParser.BROWSER.MAJOR]];
-    var parser = new UAParser(uaString, {browser: ownBrowser});
-    assert.equal(parser.getBrowser().name, 'ownbrowser');
-    assert.equal(parser.getBrowser().major, '1');
-    assert.equal(parser.getBrowser().version, '1.3');
-});
-
 for (var i in methods) {
     describe(methods[i]['title'], function () {
         for (var j in methods[i]['list']) {
@@ -72,3 +64,28 @@ for (var i in methods) {
         }
     });
 }
+
+describe('Extending Regex', function () {
+    var uaString = 'Mozilla/5.0 MyOwnBrowser/1.3';
+    var myOwnBrowser = [[/(myownbrowser)\/((\d+)?[\w\.]+)/i], [UAParser.BROWSER.NAME, UAParser.BROWSER.VERSION, UAParser.BROWSER.MAJOR]];
+    var parser = new UAParser(uaString, {browser: myOwnBrowser});
+    assert.equal(parser.getBrowser().name, 'MyOwnBrowser');
+    assert.equal(parser.getBrowser().version, '1.3');
+    assert.equal(parser.getBrowser().major, '1');
+});
+
+describe('Using Require.js', function () {
+    it('should loaded automatically', function(done) {
+        requirejs.config({
+            baseUrl : 'dist',
+            paths   : {
+                'ua-parser-js' : 'ua-parser.min'
+            }
+        });
+        requirejs(['ua-parser-js'], function(ua) {
+            var parser = new ua('Dillo/1.0');
+            assert.deepEqual(parser.getBrowser().name, 'Dillo');
+            done();
+        });
+    });
+});
